RFC - Fake Cast mode
In debug mode, we can force a "fake" Cast mode to make it easier to
test the receiver app.  This can be triggered by opening the demo app
and Cast receiver app in two different tabs or windows of the same
browser instance, then appending ";fakeCast" to the end of both URLs
and reloading the tabs.

Change-Id: I070c7148c87bf6e45cda9b2ae8f314ff37dbb683